# docs/swagger.yaml
basePath: /api/v1/
definitions:
...
  entity.Address:
    properties:
      complement:
        type: string
      number:
        type: string
      street:
        type: string
      zip:
        type: string
    required:
    - number
    - street
    - zip
    type: object
  entity.Car:
    properties:
      capacity:
        type: integer
      name:
        type: string
      year:
        type: string
    required:
    - capacity
    - name
    - year
    type: object
  entity.Driver:
    properties:
      accessibility:
        type: boolean
      address:
        $ref: '#/definitions/entity.Address'
      amount:
        type: number
      biography:
        type: string
      car:
        $ref: '#/definitions/entity.Car'
      city:
        type: string
      cnh:
        type: string
      cpf:
        type: string
      created_at:
        type: string
      descriptions:
        type: string
      email:
        type: string
      id:
        type: integer
      municipal_record:
        type: string
      name:
        type: string
      password:
        type: string
      phone:
        example: +55 11 123456789
        type: string
      pix_key:
        type: string
      profile_image:
        type: string
      qrcode:
        type: string
      schedule:
        type: string
      seats:
        $ref: '#/definitions/entity.Seats'
      states:
        type: string
      updated_at:
        type: string
    required:
    - address
    - amount
    - car
    - cnh
    - email
    - municipal_record
    - name
    - phone
    type: object
  entity.Invite:
    properties:
      accepted_at:
        type: string
      created_at:
        type: string
      driver:
        $ref: '#/definitions/entity.Driver'
      driver_id:
        type: string
      id:
        type: integer
      school:
        $ref: '#/definitions/entity.School'
      school_id:
        type: string
      status:
        type: string
      updated_at:
        type: string
    type: object
...
  entity.School:
    properties:
      address:
        $ref: '#/definitions/entity.Address'
      city:
        type: string
      cnpj:
        type: string
      created_at:
        type: string
      email:
        type: string
      id:
        type: integer
      name:
        type: string
      password:
        type: string
      phone:
        example: +55 11 123456789
        type: string
      profile_image:
        type: string
      states:
        type: string
      updated_at:
        type: string
    required:
    - address
    - cnpj
    - email
    - name
    - password
    - phone
    type: object
  entity.Seats:
    properties:
      afternoon:
        type: integer
      morning:
        type: integer
      night:
        type: integer
      remaining:
        type: integer
    type: object
...
  value.DriverListContracts:
    properties:
      amount:
        type: number
      created_at:
        type: integer
      expire_at:
        type: integer
      id:
        type: integer
      kid:
        $ref: '#/definitions/value.GetKidContract'
      record:
        type: string
      responsible:
        $ref: '#/definitions/value.GetParentContract'
      school:
        $ref: '#/definitions/value.GetSchoolContract'
      status:
        type: string
    type: object
  value.DriverListInvite:
    properties:
      address:
        type: string
      email:
        type: string
      name:
        type: string
      phone:
        type: string
      profile_image:
        type: string
    type: object
...
  value.GetDriver:
    properties:
      accessibility:
        type: boolean
      amount:
        type: number
      biography:
        type: string
      car:
        type: string
      city:
        type: string
      created_at:
        type: string
      descriptions:
        type: string
      email:
        type: string
      gallery:
        additionalProperties:
          type: string
        type: object
      id:
        type: integer
      name:
        type: string
      phone:
        type: string
      profile_image:
        type: string
      qrcode:
        type: string
      states:
        type: string
    type: object
...
  value.GetKidContract:
    properties:
      id:
        type: integer
      name:
        type: string
      period:
        type: string
      profile_image:
        type: string
    type: object
  value.GetParentContract:
    properties:
      address:
        type: string
      email:
        type: string
      id:
        type: integer
      name:
        type: string
      phone:
        type: string
      profile_image:
        type: string
    type: object
...
  value.GetSchoolContract:
    properties:
      address:
        type: string
      id:
        type: integer
      name:
        type: string
      phone:
        type: string
      profile_image:
        type: string
    type: object
  value.GetTempContracts:
    properties:
      created_at:
        description: epoch time
        type: integer
      driver_assigned_at:
        description: epoch time
        type: integer
      driver_cnh:
        type: string
      expires_at:
        description: epoch time
        type: integer
      id:
        type: integer
      kid_rg:
        type: string
      responsible_cpf:
        type: string
      responsibles_assigned_at:
        description: epoch time
        type: integer
      school_cnpj:
        type: string
      signature_request_id:
        type: string
      signing_url:
        type: string
      status:
        type: string
      uuid:
        type: string
    type: object
...
  value.SchoolListInvite:
    properties:
      email:
        type: string
      id:
        type: integer
      name:
        type: string
      phone:
        type: string
      profile_image:
        type: string
    type: object
...
host: localhost:9999
info:
  contact: {}
  description: Venture Backend Major API
  termsOfService: http://swagger.io/terms/
  title: Venture API
  version: "0.1"
paths:
...
  /contract/driver/{cnh}:
    get:
      description: Retorna todos os contratos associados ao CNH do motorista
      parameters:
      - description: CNH do motorista
        in: path
        name: cnh
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              items:
                $ref: '#/definitions/value.DriverListContracts'
              type: array
            type: array
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Lista contratos de um motorista
      tags:
      - Contracts
...
  /driver:
    post:
      consumes:
      - application/json
      parameters:
      - description: Dados do motorista
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/entity.Driver'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/value.GetDriver'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Cria um novo motorista
      tags:
      - Drivers
  /driver/{cnh}:
    delete:
      parameters:
      - description: CNH do motorista
        in: path
        name: cnh
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Deleta motorista
      tags:
      - Drivers
    get:
      description: Retorna os dados de um motorista pelo CNH
      parameters:
      - description: CNH do motorista
        in: path
        name: cnh
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/value.GetDriver'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Busca motorista
      tags:
      - Drivers
    patch:
      consumes:
      - application/json
      parameters:
      - description: CNH do motorista
        in: path
        name: cnh
        required: true
        type: string
      - description: Campos a serem atualizados
        in: body
        name: body
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "204":
          description: No Content
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Atualiza dados do motorista
      tags:
      - Drivers
  /driver/invites/{cnh}:
    get:
      description: Retorna todos os convites enviados para o motorista
      parameters:
      - description: CNH do motorista
        in: path
        name: cnh
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              items:
                $ref: '#/definitions/value.DriverListInvite'
              type: array
            type: array
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Lista convites do motorista
      tags:
      - Invites
  /driver/login:
    post:
      consumes:
      - application/json
      parameters:
      - description: Credenciais de login (email e senha)
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/entity.Driver'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Login de motorista
      tags:
      - Drivers
...
  /invite:
    post:
      consumes:
      - application/json
      description: Cria um convite entre escola e motorista
      parameters:
      - description: Dados do convite
        in: body
        name: invite
        required: true
        schema:
          $ref: '#/definitions/entity.Invite'
      produces:
      - application/json
      responses:
        "201":
          description: Created
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Envia um convite
      tags:
      - Invites
  /invite/{id}/accept:
    patch:
      description: Aceita um convite existente
      parameters:
      - description: ID do convite
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Aceita convite
      tags:
      - Invites
  /invite/{id}/decline:
    delete:
      description: Recusa e deleta um convite existente
      parameters:
      - description: ID do convite
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Recusa convite
      tags:
      - Invites
...
  /school/invites/{cnpj}:
    get:
      description: Retorna todos os convites enviados pela escola
      parameters:
      - description: CNPJ da escola
        in: path
        name: cnpj
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              items:
                $ref: '#/definitions/value.SchoolListInvite'
              type: array
            type: array
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Lista convites da escola
      tags:
      - Invites
...
  /temporary-contract/driver/{cnh}:
    get:
      description: Retorna todos os contratos temporários associados ao CNH do motorista
      parameters:
      - description: CNH do motorista
        in: path
        name: cnh
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              items:
                $ref: '#/definitions/value.GetTempContracts'
              type: array
            type: array
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Busca contratos temporários de um motorista
      tags:
      - Temporary Contracts
...
swagger: "2.0"